frontend/storm_wiki/requirements.txt
# storm
knowledge_storm==0.2.3
openai==1.36.1 # openai version for knowledge_storm 0.2.3
ollama==0.3.0

# streamlit
streamlit==1.36.0
streamlit-card
markdown
unidecode
extra-streamlit-components==0.1.60
streamlit_extras
deprecation==2.1.0
st-pages==0.4.5
streamlit-float
streamlit-option-menu

# anthropic
anthropic==0.32.0

# duckduckgo
langchain-community==0.2.10
duckduckgo-search==6.2.3

# tracing
python-dotenv
arize-phoenix==4.12.0

# tests
pytest==8.3.1
flake8==7.1.0

